+++ Frontend/src/pages/about.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import BlogBanner from '../sections/BlogBanner'
+import PageBanner from '../components/PageBanner';
 import BlogSection from '../sections/BlogSection'
 import FeaturesSection from '../sections/FeaturesSection'
 import { usePageMeta } from '../shared/lib/usePageMeta.js';
@@ -9,7 +9,7 @@ function About() {
   usePageMeta('/about');
   return (
     <div>
-      <BlogBanner />
+      <PageBanner title="Blog" trail={[{ label: 'Home', to: '/' }, { label: 'Blog' }]} />
       <BlogSection />
       <FeaturesSection />
     </div>